Feature: Added support for opening Git root in VS Code (#13498)
