+++ src/Files.App/UserControls/StatusBarControl.xaml
@@ -163,62 +163,48 @@
 			Orientation="Horizontal"
 			Spacing="4">
 
-			<!--  Open in VS Code Button  -->
+			<!--  Open in IDE Button  -->
 			<Button
-				x:Name="OpenInVSCodeButton"
+				x:Name="OpenInIDEButton"
 				Height="24"
 				Padding="8,0,8,0"
 				VerticalAlignment="Center"
-				x:Load="{x:Bind converters:MultiBooleanConverter.AndNotConvert(Commands.OpenInVSCode.IsExecutable, Commands.OpenInVS.IsExecutable), Mode=OneWay}"
+				x:Load="{x:Bind Commands.OpenInVSCode.IsExecutable, Mode=OneWay}"
 				Background="Transparent"
-				BorderBrush="Transparent"
-				Command="{x:Bind Commands.OpenInVSCode}"
-				ToolTipService.ToolTip="{x:Bind Commands.OpenInVSCode.LabelWithHotKey, Mode=OneWay}">
+				BorderBrush="Transparent">
 				<Button.Content>
 					<StackPanel Orientation="Horizontal" Spacing="8">
 						<usercontrols:OpacityIcon
 							Width="16"
 							Height="16"
 							Style="{StaticResource ColorIconOpen}" />
-						<TextBlock Text="{x:Bind Commands.OpenInVSCode.Label, Mode=OneWay}" />
+						<TextBlock Text="{helpers:ResourceString Name=Open}" />
 					</StackPanel>
 				</Button.Content>
+
+				<Button.Flyout>
+					<MenuFlyout>
+						<MenuFlyoutItem
+							x:Name="OpenFolderInVSCodeButton"
+							x:Load="{x:Bind Commands.OpenInVSCode.IsExecutable, Mode=OneWay}"
+							Command="{x:Bind Commands.OpenInVSCode}"
+							KeyboardAcceleratorTextOverride="{x:Bind Commands.OpenInVSCode.HotKeys, Mode=OneWay}"
+							Text="{x:Bind Commands.OpenInVSCode.Label, Mode=OneWay}" />
+						<MenuFlyoutItem
+							x:Name="OpenRepoInVSCodeButton"
+							x:Load="{x:Bind Commands.OpenRepoInVSCode.IsExecutable, Mode=OneWay}"
+							Command="{x:Bind Commands.OpenRepoInVSCode}"
+							KeyboardAcceleratorTextOverride="{x:Bind Commands.OpenRepoInVSCode.HotKeys, Mode=OneWay}"
+							Text="{x:Bind Commands.OpenRepoInVSCode.Label, Mode=OneWay}" />
+					</MenuFlyout>
+				</Button.Flyout>
 			</Button>
 			<!--  Divider  -->
 			<Border
 				x:Name="VSCodeDivider"
 				Width="1"
 				Height="18"
-				x:Load="{x:Bind converters:MultiBooleanConverter.AndNotConvert(Commands.OpenInVSCode.IsExecutable, Commands.OpenInVS.IsExecutable), Mode=OneWay}"
-				Background="{ThemeResource DividerStrokeColorDefaultBrush}" />
-
-			<!--  Open in VS Button  -->
-			<Button
-				x:Name="OpenInVSButton"
-				Height="24"
-				Padding="8,0,8,0"
-				VerticalAlignment="Center"
-				x:Load="{x:Bind Commands.OpenInVS.IsExecutable, Mode=OneWay}"
-				Background="Transparent"
-				BorderBrush="Transparent"
-				Command="{x:Bind Commands.OpenInVS}"
-				ToolTipService.ToolTip="{x:Bind Commands.OpenInVS.LabelWithHotKey, Mode=OneWay}">
-				<Button.Content>
-					<StackPanel Orientation="Horizontal" Spacing="8">
-						<usercontrols:OpacityIcon
-							Width="16"
-							Height="16"
-							Style="{StaticResource ColorIconOpen}" />
-						<TextBlock Text="{x:Bind Commands.OpenInVS.Label, Mode=OneWay}" />
-					</StackPanel>
-				</Button.Content>
-			</Button>
-			<!--  Divider  -->
-			<Border
-				x:Name="VSDivider"
-				Width="1"
-				Height="18"
-				x:Load="{x:Bind Commands.OpenInVS.IsExecutable, Mode=OneWay}"
+				x:Load="{x:Bind Commands.OpenInVSCode.IsExecutable, Mode=OneWay}"
 				Background="{ThemeResource DividerStrokeColorDefaultBrush}" />
 
 			<Button
